double

+++ callmainwindow.h
@@ -18,13 +18,13 @@ public:
     QLCDNumber* lcdNumber = nullptr;
     CallMainWindow(QWidget* parent = nullptr) : QMainWindow(parent) {}
 
-    bool valueError (int& value1, int& value2)
+    bool valueError (double& value1, double& value2)
     {
         bool correctValue1 = true;
         bool correctValue2 = true;
 
-        value1 = lineEdit_2->text().toInt(&correctValue1);
-        value2 = lineEdit->text().toInt(&correctValue2);
+        value1 = lineEdit_2->text().toDouble(&correctValue1);
+        value2 = lineEdit->text().toDouble(&correctValue2);
 
         return (correctValue1 && correctValue2);
     }
@@ -33,8 +33,8 @@ public:
 public slots:
     void addition()
     {
-        int value1 {0};
-        int value2 {0};
+        double value1 {0};
+        double value2 {0};
 
         if(valueError(value1,value2))
             lcdNumber->display(value1 + value2);
@@ -44,8 +44,8 @@ public slots:
 
     void subtraction()
     {
-        int value1 {0};
-        int value2 {0};
+        double value1 {0};
+        double value2 {0};
 
         if(valueError(value1,value2))
             lcdNumber->display(value1 - value2);
@@ -55,8 +55,8 @@ public slots:
 
     void multiplication()
     {
-        int value1 {0};
-        int value2 {0};
+        double value1 {0};
+        double value2 {0};
 
         if(valueError(value1,value2))
             lcdNumber->display(value1 * value2);
@@ -66,8 +66,8 @@ public slots:
 
     void division()
     {
-        int value1 {0};
-        int value2 {0};
+        double value1 {0};
+        double value2 {0};
 
         if(valueError(value1,value2))
             lcdNumber->display(value1 / value2);
